+++ modules/apps/foundation/portal/portal-messaging/src/main/java/com/liferay/portal/messaging/internal/DefaultMessageBus.java
@@ -30,7 +30,7 @@ import com.liferay.portal.kernel.resiliency.spi.SPIUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.messaging.configuration.DestinationWorkerConfiguration;
+import com.liferay.portal.messaging.internal.configuration.DestinationWorkerConfiguration;
 
 import java.util.ArrayList;
 import java.util.Collection;
