+++ modules/apps/foundation/portal/portal-messaging/src/main/java/com/liferay/portal/messaging/internal/DestinationPrototype.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.messaging;
+package com.liferay.portal.messaging.internal;
 
 import com.liferay.portal.kernel.messaging.Destination;
 import com.liferay.portal.kernel.messaging.DestinationConfiguration;
