+++ modules/apps/foundation/portal/portal-messaging/src/main/java/com/liferay/portal/messaging/internal/configuration/DestinationWorkerConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.messaging.configuration;
+package com.liferay.portal.messaging.internal.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
